requirements.txt
click>=8.0.0
pyyaml>=6.0.0
rich>=10.0.0 